+++ modules.tf
@@ -74,3 +74,41 @@ module "keyvault" {
 
   network_acls = var.keyvault_network_acls
 }
+
+module "monitoring-function" {
+  source = "./modules/monitoring_function"
+
+  count = var.monitoring_function_enabled ? 1 : 0
+
+  client_name         = var.client_name
+  environment         = var.environment
+  location            = var.location
+  location_short      = var.location_short
+  resource_group_name = coalesce(var.keyvault_resource_group_name, var.resource_group_name)
+  stack               = var.stack
+
+  zip_package_path           = var.monitoring_function_zip_package_path
+  metrics_extra_dimensions   = var.monitoring_function_metrics_extra_dimensions
+  extra_application_settings = var.monitoring_function_extra_application_settings
+
+  log_analytics_workspace_guid = module.logs.log_analytics_workspace_guid
+  splunk_token                 = var.monitoring_function_splunk_token
+
+  logs_destinations_ids = [
+    module.logs.log_analytics_workspace_id,
+    module.logs.logs_storage_account_id,
+  ]
+
+  logs_retention_days     = var.log_analytics_workspace_retention_in_days
+  logs_categories         = var.monitoring_function_logs_categories
+  logs_metrics_categories = var.monitoring_function_logs_metrics_categories
+}
+
+resource "azurerm_role_assignment" "function_workspace" {
+  count = var.monitoring_function_enabled && var.monitoring_function_assign_role_on_workspace ? 1 : 0
+
+  principal_id = module.monitoring-function[0].function_app_identity["principal_id"]
+  scope        = module.logs.log_analytics_workspace_id
+
+  role_definition_name = "Log Analytics Reader"
+}