+++ modules/monitoring_function/r-queries.tf
@@ -1,0 +1,15 @@
+resource "azurerm_storage_table" "queries" {
+  name                 = "LogQueries"
+  storage_account_name = module.function.storage_account_name
+}
+
+resource "azurerm_storage_table_entity" "queries" {
+  for_each = local.log_queries
+
+  storage_account_name = module.function.storage_account_name
+  table_name           = azurerm_storage_table.queries.name
+
+  partition_key = "LogQuery"
+  row_key       = each.key
+  entity        = each.value
+}