+++ modules/monitoring_function/versions.tf
@@ -1,9 +1,9 @@
 terraform {
-  required_version = ">= 0.13.0"
+  required_version = "> 0.12.26"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.48"
+      version = ">= 2.21"
     }
   }
 }